+++ lib/src/commonMain/kotlin/co/touchlab/kjwt/builder/JwtBuilder.kt
@@ -3,6 +3,10 @@ package co.touchlab.kjwt.builder
 import co.touchlab.kjwt.cryptography.SimpleKey
 import co.touchlab.kjwt.cryptography.processors.CryptographyKotlinEncryptionProcessor
 import co.touchlab.kjwt.cryptography.processors.CryptographyKotlinIntegrityProcessor
+import co.touchlab.kjwt.cryptography.registry.CryptographyKotlinJwtKeyRegistry
+import co.touchlab.kjwt.cryptography.registry.EncryptionKey
+import co.touchlab.kjwt.cryptography.registry.SigningKey
+import co.touchlab.kjwt.cryptography.registry.SigningKey.Identifier
 import co.touchlab.kjwt.internal.encodeBase64Url
 import co.touchlab.kjwt.internal.encodeToBase64Url
 import co.touchlab.kjwt.model.JwtHeader
@@ -11,11 +15,7 @@ import co.touchlab.kjwt.model.JwtPayload
 import co.touchlab.kjwt.model.algorithm.EncryptionAlgorithm
 import co.touchlab.kjwt.model.algorithm.EncryptionContentAlgorithm
 import co.touchlab.kjwt.model.algorithm.SigningAlgorithm
-import co.touchlab.kjwt.model.registry.CryptographyKotlinJwtKeyRegistry
-import co.touchlab.kjwt.model.registry.EncryptionKey
 import co.touchlab.kjwt.model.registry.JwtKeyRegistry
-import co.touchlab.kjwt.model.registry.SigningKey
-import co.touchlab.kjwt.model.registry.SigningKey.Identifier
 import co.touchlab.kjwt.processor.JweProcessor
 import co.touchlab.kjwt.processor.JwsProcessor
 import dev.whyoleg.cryptography.materials.key.Key
