+++ lib/src/commonTest/kotlin/co/touchlab/kjwt/JweEncodeTest.kt
@@ -5,7 +5,7 @@ package co.touchlab.kjwt
 import co.touchlab.kjwt.ext.subjectOrNull
 import co.touchlab.kjwt.model.algorithm.EncryptionAlgorithm
 import co.touchlab.kjwt.model.algorithm.EncryptionContentAlgorithm
-import co.touchlab.kjwt.model.registry.EncryptionKey
+import co.touchlab.kjwt.cryptography.registry.EncryptionKey
 import dev.whyoleg.cryptography.algorithms.RSA
 import io.kotest.core.spec.style.FunSpec
 import kotlin.test.assertEquals
