+++ lib/src/commonTest/kotlin/co/touchlab/kjwt/JwtKeyRegistryTest.kt
@@ -4,9 +4,9 @@ import co.touchlab.kjwt.ext.subjectOrNull
 import co.touchlab.kjwt.model.algorithm.EncryptionAlgorithm
 import co.touchlab.kjwt.model.algorithm.EncryptionContentAlgorithm
 import co.touchlab.kjwt.model.algorithm.SigningAlgorithm
-import co.touchlab.kjwt.model.registry.EncryptionKey
-import co.touchlab.kjwt.model.registry.CryptographyKotlinJwtKeyRegistry
-import co.touchlab.kjwt.model.registry.SigningKey
+import co.touchlab.kjwt.cryptography.registry.EncryptionKey
+import co.touchlab.kjwt.cryptography.registry.CryptographyKotlinJwtKeyRegistry
+import co.touchlab.kjwt.cryptography.registry.SigningKey
 import io.kotest.core.spec.style.FunSpec
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
